Record punishments in LadderInfo instead of per user
For easier management of course!
